requirements.txt
# packages installed by pip.
absl-py==2.1.0
accelerate==0.34.2
aenum==3.1.15
aiohappyeyeballs==2.4.2
aiohttp==3.10.6
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.2.0
bitsandbytes==0.42.0
black==23.11.0
cachetools==5.5.0
cairocffi==1.7.1
cairosvg==2.7.1
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.3.2
colorlog==6.8.2
comm==0.2.2
contourpy==1.3.0
cssselect2==0.7.0
cycler==0.12.1
datasets==2.15.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
distlib==0.3.8
docker-pycreds==0.4.0
dreamsim==0.1.3
einops==0.7.0
ema-pytorch==0.5.3
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2023.10.0
ftfy==6.2.3
gast==0.6.0
gitdb==4.0.11
gitpython==3.1.43
google-auth==2.35.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.66.1
h5py==3.12.1
haishoku==1.1.8
huggingface-hub==0.25.1
hydra-colorlog==1.2.0
hydra-core==1.3.2
identify==2.6.1
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
ipykernel==6.29.4
ipython==8.18.1
ipywidgets==8.1.1
isort==5.12.0
jedi==0.19.1
jinja2==3.1.4
joblib==1.4.2
jupyter-client==8.6.1
jupyter-core==5.7.2
jupyterlab-widgets==3.0.13
keras==2.14.0
kiwisolver==1.4.7
libclang==18.1.1
lightly==1.5.13
lightly-utils==0.0.2
lightning-utilities==0.11.7
markdown==3.7
markupsafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
ml-dtypes==0.2.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.15
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nodeenv==1.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
open-clip-torch==2.26.1
opencv-python==4.8.1.78
opt-einsum==3.4.0
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
peft==0.1.0
pexpect==4.9.0
pillow==10.4.0
prdc==0.2
pre-commit==3.5.0
prompt-toolkit==3.0.48
protobuf==4.25.5
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==17.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1-modules==0.4.1
pybind11==2.13.6
# pycairo==1.26.1
pycparser==2.22
pydantic==1.10.18
pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytorch-fid==0.3.0
pytorch-lightning==2.4.0
pytz==2024.2
pyyaml==6.0.2
pyzmq==26.2.0
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.11.1
seaborn==0.13.0
sentry-sdk==2.12.0
setproctitle==1.3.3
six==1.16.0
skia-python==87.5
smmap==5.0.1
stack-data==0.6.3
sympy==1.13.3
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
thop==0.1.1-2209072238
threadpoolctl==3.5.0
timm==1.0.9
tinycss==0.4
tinycss2==1.3.0
tokenizers==0.20.0
# torch==2.1.1+cu121
# torchaudio==2.1.1+cu121
# torchmetrics==1.4.2
# torchvision==0.16.1+cu121
# use pip install torch==2.1.1 torchvision==0.16.1 torchaudio==2.1.1 --index-url https://download.pytorch.org/whl/cu121
torchviz==0.0.2
tornado==6.4.1
tqdm==4.66.1
traitlets==5.14.3
transformers==4.45.1
triton==2.1.0
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.25.0
wandb==0.17.2
wcwidth==0.2.13
webencodings==0.5.1
werkzeug==3.0.4
widgetsnbextension==4.0.13
wrapt==1.14.1
xxhash==3.5.0
yarl==1.13.0
zipp==3.20.2
scikit-image==0.24.0